split into short and long
