+++ test/minikube/minikube_tls_rotation_test.go
@@ -1,3 +1,5 @@
+// +build long
+
 package minikube
 
 import (
